hubor/requirements.txt
django==3.0.6
djangorestframework
markdown
django-filter
channels